requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==21.2.0
autopep8==1.5.4
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
ConfigArgParse==1.2.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
Flask==1.1.2
Flask-BasicAuth==0.2.0
fonttools==4.28.5
gevent==20.9.0
geventhttpclient==1.4.4
gitdb==4.0.9
GitPython==3.1.24
google-auth==1.30.1
google-auth-oauthlib==0.4.4
greenlet==0.4.17
gspread==3.7.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.2
joblib==1.1.0
jsonschema==4.3.2
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
locust==1.4.1
MarkupSafe==1.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.1
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.1.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
progress==1.6
progress-bar==8
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
psutil==5.7.3
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.10.0
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
six==1.15.0
smmap==5.0.0
soupsieve==2.3.1
streamlit==1.3.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tradingview-ta==3.2.9
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.2
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==1.0.1
widgetsnbextension==3.5.2
zope.event==4.5.0
zope.interface==5.2.0
